:sparkles: Override mj-wrapper validation rule & add validation parameter (#1)
This PR adds a new query parameter `validation` to specify the validation level you need. By default it's `strict` but can take the `soft` and `skip` value.

Plus, by default we want to allow the `mj-wrapper` element be nested with another `mj-wrapper` element. To do this, we override the default validation for the `mj-wrapper` component. The reason behind this change is related to our specific use-case.